check if order should be returned in amocrm (#1957)

+++ src/amocrm/services/orders/order_returner.py
@@ -8,13 +8,14 @@ from orders.models import Order
 
 @dataclass
 class AmoCRMOrderReturner(BaseService):
-    """
-    Updates amocrm_lead for given order and deletes transaction
-    """
+    """Updates amocrm_lead for given order and deletes transaction"""
 
     order: Order
 
     def act(self) -> None:
+        if self.order.amocrm_transaction is None:  # order is not pushed to amocrm as purchased
+            return
+
         AmoCRMLead(order=self.order).update(status="closed")
         AmoCRMTransaction(order=self.order).delete()
-        self.order.amocrm_transaction.delete()  # type: ignore
+        self.order.amocrm_transaction.delete()
